[refactor/project-structure] Refactor piu' grande fatto finora: ristrutturata l'intera struttura del progetto; eliminate funzioni get32/put32 in quanto gia' presenti mmio_write/read (che sono migliori sotto ogni aspetto); rinominati irq.c --> controller.c, irq.h --> controller.h, irq_setup.S --> vector.S, cosi' e' piu' untuitivo
