--- drivers/utils.h
@@ -1,9 +1,0 @@
-#ifndef	_UTILS_H
-#define	_UTILS_H
-
-extern int get_el ( void );
-extern void delay ( unsigned long);
-extern void put32 ( unsigned long, unsigned int );
-extern unsigned int get32 ( unsigned long );
-
-#endif
