+++ libs/utils.h
@@ -1,0 +1,8 @@
+#ifndef _UTILS_H
+#define _UTILS_H
+
+extern int get_el(void);
+extern void delay(unsigned long);
+
+#endif
+
